Support floats #3

+++ src/Data/Amount.php
@@ -7,7 +7,7 @@ namespace Icepay\Payment\Data;
 class Amount
 {
     public function __construct(
-        public readonly int $value,
+        public readonly float $value,
         public readonly string $currency
     ) {}
 }
